Fix text values returning float instead of int

+++ gui.py
@@ -48,12 +48,12 @@ class Window:
 
         # Create label
         self.label = Label(self.root, text=None)
-        self.label.configure(font=('Calibri', 18*self.wscale))
+        self.label.configure(font=('Calibri', int(18*self.wscale)))
         self.label.grid(row=2, rowspan=2, sticky='NWSE')
 
         # Create button
         self.button = Button(self.root, text="Hear about Pluto?", command=self.update)
-        self.button.configure(fg='#191970', activeforeground='white', bd=0, font=('Calibri', 16*self.wscale))
+        self.button.configure(fg='#191970', activeforeground='white', bd=0, font=('Calibri', int(16*self.wscale)))
         self.button.configure(highlightthickness=0, highlightbackground='#708090')
         # self.button.grid(row=4, pady=14, sticky='NWSE')
         self.button.grid(row=4, pady=0.5*self.hscale, sticky='NWSE')
@@ -89,7 +89,7 @@ class Window:
             im = Image.open(self.get_image())
 
         # im = im.resize((403, 403), Image.ANTIALIAS)
-        im = im.resize((403*self.wscale, 403*self.hscale), Image.ANTIALIAS)
+        im = im.resize((int(403*self.wscale), int(403*self.hscale)), Image.ANTIALIAS)
         print(self.wscale, self.hscale)
 
         im = ImageTk.PhotoImage(im)
